requirements.txt
attrs==23.2.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
coverage==7.4.4
distlib==0.3.8
filelock==3.13.3
h11==0.14.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
PySocks==1.7.1
pytest==8.1.1
pytest-cov==5.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
ruff==0.3.5
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
types-requests==2.31.0.20240403
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
wsproto==1.2.0
